requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via flask
colorama==0.4.6
    # via click
deepgram-sdk==5.3.2
    # via bilingual-transcription-demo (pyproject.toml)
flask==3.1.2
    # via bilingual-transcription-demo (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via deepgram-sdk
idna==3.11
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
pydantic==2.12.5
    # via deepgram-sdk
pydantic-core==2.41.5
    # via
    #   deepgram-sdk
    #   pydantic
pydub==0.25.1
    # via bilingual-transcription-demo (pyproject.toml)
python-dotenv==1.2.1
    # via bilingual-transcription-demo (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   anyio
    #   deepgram-sdk
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
websockets==16.0
    # via deepgram-sdk
werkzeug==3.1.5
    # via flask
